refactor: move interfaces to one dir
