+++ src/interfaces/slices.ts
@@ -1,0 +1,16 @@
+export interface ImageData {
+  url: string;
+  author: string;
+  id: string;
+}
+
+export interface FilterValue {
+  author: string;
+  authors: string[];
+}
+
+export interface UserInterface {
+    value: string | null;
+    name: string;
+    error: string | null;
+  }
