+++ src/utils/canvasUtils.ts
@@ -40,10 +40,9 @@ export const drawCircle = (
   coords: Coords | null,
 ) => {
   if (!coords) return;
-  const x =  Math.abs(coords.start.x - coords.end.x)
-  const y =  Math.abs(coords.start.y - coords.end.y)
-  const rad =
-  x > y ? x : y
+  const x = Math.abs(coords.start.x - coords.end.x);
+  const y = Math.abs(coords.start.y - coords.end.y);
+  const rad = x > y ? x : y;
   context.beginPath();
   context.arc(coords.start.x, coords.start.y, rad, 0, 2 * Math.PI);
   context.stroke();
